CaseView merged with CaseForm
